requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-browser-reload
    #   django-simple-history
black==24.4.1
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
contourpy==1.2.1
    # via matplotlib
coverage==7.5.0
    # via -r requirements.in
crispy-bootstrap5==2024.2
    # via -r requirements.in
cryptography==42.0.5
    # via pyopenssl
cycler==0.12.1
    # via matplotlib
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
dj-database-url==2.1.0
    # via -r requirements.in
dj-user-login-history==1.0.5
    # via -r requirements.in
django==4.2.11
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   dj-database-url
    #   dj-user-login-history
    #   django-admin-autocomplete-filter
    #   django-admin-logs
    #   django-appconf
    #   django-axes
    #   django-browser-reload
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-import-export
    #   django-mailjet
    #   django-recaptcha
    #   django-select2
    #   django-user-visit
    #   djangorestframework
django-admin-autocomplete-filter==0.7.1
    # via -r requirements.in
django-admin-logs==1.2.0
    # via -r requirements.in
django-admin-rangefilter==0.12.4
    # via -r requirements.in
django-appconf==1.0.6
    # via django-select2
django-auto-logout==0.5.1
    # via -r requirements.in
django-axes==6.4.0
    # via -r requirements.in
django-browser-reload==1.12.1
    # via -r requirements.in
django-crispy-forms==2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-debug-toolbar==4.3.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
django-grappelli==4.0.1
    # via -r requirements.in
django-import-export==3.3.8
    # via -r requirements.in
django-mailjet==0.3.1
    # via -r requirements.in
django-recaptcha==4.0.0
    # via -r requirements.in
django-select2==8.1.2
    # via -r requirements.in
django-simple-history==3.5.0
    # via -r requirements.in
django-user-visit==2.0
    # via -r requirements.in
django-widget-tweaks==1.5.0
    # via -r requirements.in
djangoql==0.18.1
    # via -r requirements.in
djangorestframework==3.15.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
faker==12.0.1
    # via mixer
fonttools==4.51.0
    # via matplotlib
gunicorn==21.0.1
    # via -r requirements.in
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
kiwisolver==1.4.5
    # via matplotlib
mailjet-rest==1.3.4
    # via django-mailjet
markdown==3.6
    # via -r requirements.in
markuppy==1.14
    # via tablib
markupsafe==2.1.5
    # via werkzeug
matplotlib==3.8.4
    # via -r requirements.in
mixer==7.2.2
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   pandas
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via
    #   -r requirements.in
    #   tablib
packaging==24.0
    # via
    #   black
    #   gunicorn
    #   matplotlib
    #   pytest
    #   pytest-sugar
pandas==2.2.2
    # via -r requirements.in
pathspec==0.12.1
    # via black
pillow==10.3.0
    # via matplotlib
platformdirs==4.2.1
    # via black
pluggy==1.5.0
    # via pytest
ply==3.11
    # via djangoql
psycopg[binary]==3.1.18
    # via -r requirements.in
psycopg-binary==3.1.18
    # via psycopg
pycparser==2.22
    # via cffi
pydotplus==2.0.2
    # via -r requirements.in
pyopenssl==24.1.0
    # via -r requirements.in
pyparsing==3.1.2
    # via
    #   matplotlib
    #   pydotplus
pytest==8.1.1
    # via
    #   pytest-django
    #   pytest-mock
    #   pytest-sugar
pytest-django==4.8.0
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
pytest-sugar==1.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via tablib
requests==2.31.0
    # via mailjet-rest
ruff==0.4.1
    # via -r requirements.in
sentry-sdk==2.0.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via django-import-export
termcolor==2.4.0
    # via pytest-sugar
typing-extensions==4.11.0
    # via
    #   dj-database-url
    #   psycopg
tzdata==2024.1
    # via pandas
ua-parser==0.18.0
    # via user-agents
urllib3==2.2.1
    # via
    #   requests
    #   sentry-sdk
user-agents==2.2.0
    # via django-user-visit
werkzeug==3.0.2
    # via -r requirements.in
whitenoise==6.6.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlsxwriter==3.2.0
    # via -r requirements.in
xlwt==1.3.0
    # via tablib
